% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{add_cost}
\alias{add_cost}
\title{Defining costs for events and intervention}
\usage{
add_cost(.data = NULL, cost, evt, trt, cycle_l = NULL, cycle_starttime = 0)
}
\arguments{
\item{.data}{Existing cost data}

\item{cost}{Value or expression to calculate the cost estimate}

\item{evt}{Vector of events for which this cost is applicable}

\item{trt}{Vector of interventions for which this cost is applicable}

\item{cycle_l}{Cycle length; only needed if costs are calculated per cycle}

\item{cycle_starttime}{Cycle when costs start being accrued; only needed if costs are calculated per cycle}
}
\value{
A list of costs
}
\description{
Defining costs for events and intervention
}
\details{
Costs can be defined by writing expressions and objects in the cost argument whose execution will be delayed until the model runs.

This function accepts the use of pipes (\%>\%) to define multiple costs.
}
\examples{
add_cost(evt = c("start","idfs","ttot"),trt = "int",cost = cost.int*fl.int + cost.idfs)

}
