% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{add_item}
\alias{add_item}
\title{Defining parameters that may be used in model calculations}
\usage{
add_item(.data = NULL, ...)
}
\arguments{
\item{.data}{Existing data}

\item{...}{Items to define for the simulation}
}
\value{
A list of items
}
\description{
Defining parameters that may be used in model calculations
}
\details{
The functions to add/modify events/inputs use lists. Whenever several inputs/events are added or modified, it's recommended to group them within one function, as it reduces the computation cost.
So rather than use two \code{add_item} with a list of one element, it's better to group them into a single \code{add_item} with a list of two elements.
}
\examples{
add_item(fl.idfs = 0)
add_item(util_idfs = if(psa_bool){rnorm(1,0.8,0.2)} else{0.8}, util.mbc = 0.6, cost_idfs = 2500)

}
