% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_summary_f.R
\name{ceac_des}
\alias{ceac_des}
\title{Calculate the cost-effectiveness acceptability curve (CEAC) for a DES model with a PSA result}
\usage{
ceac_des(wtp, results, interventions = NULL)
}
\arguments{
\item{wtp}{Vector of length >=1 with the willingness to pay}

\item{results}{The list object returned by \code{RunSim()}}

\item{interventions}{A character vector with the names of the interventions to be used for the analysis}
}
\value{
A data frame with the CEAC results
}
\description{
Calculate the cost-effectiveness acceptability curve (CEAC) for a DES model with a PSA result
}
\examples{
\dontrun{
ceac_des(seq(from=10000,to=500000,by=10000),results)
}
}
