% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{draw.rbeta}
\alias{draw.rbeta}
\title{Draw from a beta distribution based on mean and se}
\usage{
draw.rbeta(value, se)
}
\arguments{
\item{value}{A vector of the mean values}

\item{se}{A vector of the standard errors of the means}
}
\value{
A single estimate from the beta distribution based on given parameters
}
\description{
Draw from a beta distribution based on mean and se
}
\examples{
draw.rbeta(value=0.8,se=0.2)
}
