% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{draw.rgamma}
\alias{draw.rgamma}
\title{Draw from a gamma distribution based on mean and se}
\usage{
draw.rgamma(value, se)
}
\arguments{
\item{value}{A vector of the mean values}

\item{se}{A vector of the standard errors of the means}
}
\value{
A single estimate from the gamma distribution based on given parameters
}
\description{
Draw from a gamma distribution based on mean and se
}
\examples{
draw.rgamma(value=0.8,se=0.2)

}
