% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{draw_beta}
\alias{draw_beta}
\title{Draw from a beta distribution based on mean and se}
\usage{
draw_beta(value, se, seed = NULL)
}
\arguments{
\item{value}{A vector of the mean values}

\item{se}{A vector of the standard errors of the means}

\item{seed}{An integer which will be used to set the seed for this draw. If no seed is set a random integer is used.}
}
\value{
A single estimate from the beta distribution based on given parameters
}
\description{
Draw from a beta distribution based on mean and se
}
\examples{
draw_beta(value=0.8,se=0.2)
}
