% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{draw_tte}
\alias{draw_tte}
\title{Draw a time to event from a list of parametric survival functions}
\usage{
draw_tte(
  n_chosen = 1,
  dist = "exp",
  coef1 = 1,
  coef2 = NULL,
  coef3 = NULL,
  hr = 1,
  seed = NULL
)
}
\arguments{
\item{n_chosen}{The number of observations to be drawn}

\item{dist}{The distribution; takes values 'lnorm','weibullPH','weibull','llogis','gompertz','gengamma','gamma','exp'}

\item{coef1}{First coefficient of the distribution, defined as in the coef() output on a flexsurvreg object}

\item{coef2}{Second coefficient of the distribution, defined as in the coef() output on a flexsurvreg object}

\item{coef3}{Third coefficient of the distribution, defined as in the coef() output on a flexsurvreg object}

\item{hr}{A hazard ratio}

\item{seed}{An integer which will be used to set the seed for this draw. If no seed is set a random integer is used.}
}
\value{
A vector of time to event estimates from the given parameters
}
\description{
Draw a time to event from a list of parametric survival functions
}
\examples{
draw_tte(n_chosen=1,dist='exp',coef1=1,hr=1)
}
