% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_summary_f.R
\name{evpi_des}
\alias{evpi_des}
\title{Calculate the Expected Value of Perfect Information (EVPI) for a DES model with a PSA result}
\usage{
evpi_des(wtp, results, interventions = NULL)
}
\arguments{
\item{wtp}{Vector of length >=1 with the willingness to pay}

\item{results}{The list object returned by \code{RunSim()}}

\item{interventions}{A character vector with the names of the interventions to be used for the analysis}
}
\value{
A data frame with the EVPI results
}
\description{
Calculate the Expected Value of Perfect Information (EVPI) for a DES model with a PSA result
}
\examples{
\dontrun{
evpi_des(seq(from=10000,to=500000,by=10000),results)
}
}
