% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{modify_item}
\alias{modify_item}
\title{Modify the value of existing items}
\usage{
modify_item(list_item, env_ch = NULL)
}
\arguments{
\item{list_item}{A list of items and their values or expressions}

\item{env_ch}{Environment in which to save list (should not be defined by user)}
}
\description{
Modify the value of existing items
}
\details{
The functions to add/modify events/inputs use lists. Whenever several inputs/events are added or modified, it's recommended to group them within one function, as it reduces the computation cost.
So rather than use two \code{modify_item} with a list of one element, it's better to group them into a single \code{modify_item} with a list of two elements.
}
\examples{
\dontrun{
modify_item(list(cost.idfs = 500, cost.tx = cost.tx + 4000))
}
}
