% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{new_event}
\alias{new_event}
\title{Generate new events to be added to existing vector of events}
\usage{
new_event(evt, env_ch = NULL)
}
\arguments{
\item{evt}{Event name and event time}

\item{env_ch}{Environment in which to save list (should not be defined by user)}
}
\description{
Generate new events to be added to existing vector of events
}
\details{
The functions to add/modify events/inputs use lists. Whenever several inputs/events are added or modified, it's recommended to group them within one function, as it reduces the computation cost.
So rather than use two \code{new_event} with a list of one element, it's better to group them into a single \code{new_event} with a list of two elements.
}
\examples{
\dontrun{
new_event(list("ae"=5))
new_event(list("ae"=5,"nat.death" = 100))
}
}
