% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{rresgompertz}
\alias{rresgompertz}
\title{Draw from a restricted Gompertz distribution}
\usage{
rresgompertz(n, shape, rate, lower_bound = 0, upper_bound = Inf)
}
\arguments{
\item{n}{The number of observations to be drawn}

\item{shape}{The shape parameter of the Gompertz distribution, defined as in the coef() output on a flexsurvreg object}

\item{rate}{The rate parameter of the Gompertz distribution, defined as in the coef() output on a flexsurvreg object}

\item{lower_bound}{The lower bound of the restricted distribution}

\item{upper_bound}{The upper bound of the restricted distribution}
}
\value{
Estimate(s) from the restricted Gompertz distribution based on given parameters
}
\description{
Draw from a restricted Gompertz distribution
}
\examples{
rresgompertz(1,shape=0.05,rate=0.01,lower_bound = 50)
}
