% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunEngine_debug.R
\name{RunEngine_debug}
\alias{RunEngine_debug}
\title{Run the simulation engine without using a parallel engine, facilitating debugging}
\usage{
RunEngine_debug(
  trt_list,
  common_pt_inputs = NULL,
  unique_pt_inputs = NULL,
  input_list = NULL
)
}
\arguments{
\item{trt_list}{A vector of the names of the interventions evaluated in the simulation}

\item{common_pt_inputs}{A list of inputs that change across patients but are not affected by the intervention}

\item{unique_pt_inputs}{A list of inputs that change across each intervention}

\item{input_list}{A list of all other inputs: drc, drq, psa_bool, init_event_list, evt_react_list,
uc_lists = list(util_ongoing_list,util_instant_list,util_cycle_list,cost_ongoing_list,cost_instant_list,cost_cycle_list),
input_out,ipd,trt_list,simulation,npats,n_sim}
}
\value{
A data frame with the simulation results
}
\description{
Run the simulation engine without using a parallel engine, facilitating debugging
}
\keyword{internal}
