% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_summary_f.R
\name{summary_results_det}
\alias{summary_results_det}
\title{Deterministic results for a specific treatment}
\usage{
summary_results_det(out = final_output, trt = NULL)
}
\arguments{
\item{out}{The final_output data frame from the list object returned by \code{RunSim()}}

\item{trt}{The reference treatment for calculation of incremental outcomes}
}
\value{
A dataframe with absolute costs, LYs, QALYs, and ICER and ICUR for each intervention
}
\description{
Deterministic results for a specific treatment
}
\examples{
\dontrun{
summary_results_det(results$final_output,trt="int")
}
}
