% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_summary_f.R
\name{summary_results_psa}
\alias{summary_results_psa}
\title{Summary of PSA outputs for a treatment}
\usage{
summary_results_psa(out = output_psa, trt = NULL)
}
\arguments{
\item{out}{The output_psa data frame from the list object returned by \code{RunSim()}}

\item{trt}{The reference treatment for calculation of incremental outcomes}
}
\value{
A data frame with mean and 95\% CI of absolute costs, LYs, QALYs, ICER and ICUR for each intervention from the PSA samples
}
\description{
Summary of PSA outputs for a treatment
}
\examples{
\dontrun{
summary_results_psa(results$output_psa, trt="int")
}
}
